rebase branch onto latest develop, report changes. hurt a little
