+++ plugins/hello_world.hpp
@@ -30,7 +30,6 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include "plugin_base.hpp"
 
-#include "../data_structures/json_container.hpp"
 #include "../Util/cast.hpp"
 #include "../Util/json_renderer.hpp"
