+++ routing_algorithms/routing_base.hpp
@@ -28,10 +28,10 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #ifndef ROUTING_BASE_HPP
 #define ROUTING_BASE_HPP
 
-#include "../data_structures/raw_route_data.hpp"
+#include "../data_structures/internal_route_result.hpp"
 #include "../data_structures/search_engine_data.hpp"
 #include "../data_structures/turn_instructions.hpp"
-// #include "../Util/simple_logger.hpp.h"
+// #include "../Util/simple_logger.hpp"
 
 #include <boost/assert.hpp>
